# [snapshot] Pool checkpoint buffers