--- a/pkg/backend/httpstate/diffs.go
+++ b/pkg/backend/httpstate/diffs.go
@@ -15,6 +15,7 @@
 package httpstate
 
 import (
+	"bytes"
 	"context"
 	"crypto/sha256"
 	"encoding/hex"
@@ -24,14 +25,22 @@ import (
 
 	"github.com/hexops/gotextdiff/myers"
 	"github.com/hexops/gotextdiff/span"
+	"github.com/pulumi/pulumi/pkg/v3/backend/httpstate/client"
+	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
 
 	opentracing "github.com/opentracing/opentracing-go"
 )
 
+type deployment struct {
+	raw json.RawMessage
+	buf *bytes.Buffer
+}
+
 type deploymentDiffState struct {
-	lastSavedDeployment json.RawMessage
+	lastSavedDeployment deployment
 	sequenceNumber      int
 	minimalDiffSize     int
+	buffer              *bytes.Buffer
 }
 
 type deploymentDiff struct {
@@ -52,25 +61,39 @@ func (dds *deploymentDiffState) SequenceNumber() int {
 }
 
 func (dds *deploymentDiffState) CanDiff() bool {
-	return dds.lastSavedDeployment != nil
+	return dds.lastSavedDeployment.raw != nil
+}
+
+func (dds *deploymentDiffState) MarshalDeployment(d *apitype.DeploymentV3) (deployment, error) {
+	var b *bytes.Buffer
+	if dds.buffer != nil {
+		b, dds.buffer = dds.buffer, nil
+	} else {
+		b = &bytes.Buffer{}
+	}
+
+	if err := client.MarshalUntypedDeployment(b, d); err != nil {
+		return deployment{}, err
+	}
+	return deployment{raw: json.RawMessage(b.Bytes()), buf: b}, nil
 }
 
 // Size-based heuristics trying to estimate if the diff method will be
 // worth it and take less time than sending the entire deployment.
-func (dds *deploymentDiffState) ShouldDiff(new json.RawMessage) bool {
+func (dds *deploymentDiffState) ShouldDiff(new deployment) bool {
 	if !dds.CanDiff() {
 		return false
 	}
-	if len(dds.lastSavedDeployment) < dds.minimalDiffSize {
+	if len(dds.lastSavedDeployment.raw) < dds.minimalDiffSize {
 		return false
 	}
-	if len(new) < dds.minimalDiffSize {
+	if len(new.raw) < dds.minimalDiffSize {
 		return false
 	}
 	return true
 }
 
-func (dds *deploymentDiffState) Diff(ctx context.Context, deployment json.RawMessage) (deploymentDiff, error) {
+func (dds *deploymentDiffState) Diff(ctx context.Context, deployment deployment) (deploymentDiff, error) {
 	if !dds.CanDiff() {
 		return deploymentDiff{}, fmt.Errorf("Diff() cannot be called before Saved()")
 	}
@@ -78,8 +101,8 @@ func (dds *deploymentDiffState) Diff(ctx context.Context, deployment json.RawMes
 	tracingSpan, childCtx := opentracing.StartSpanFromContext(ctx, "Diff")
 	defer tracingSpan.Finish()
 
-	before := dds.lastSavedDeployment
-	after := deployment
+	before := dds.lastSavedDeployment.raw
+	after := deployment.raw
 
 	var checkpointHash string
 	checkpointHashReady := &sync.WaitGroup{}
@@ -113,7 +136,11 @@ func (dds *deploymentDiffState) Diff(ctx context.Context, deployment json.RawMes
 }
 
 // Indicates that a deployment was just saved to the service.
-func (dds *deploymentDiffState) Saved(ctx context.Context, deployment json.RawMessage) error {
+func (dds *deploymentDiffState) Saved(ctx context.Context, deployment deployment) error {
+	if dds.lastSavedDeployment.buf != nil {
+		dds.buffer = dds.lastSavedDeployment.buf
+		dds.buffer.Reset()
+	}
 	dds.lastSavedDeployment = deployment
 	dds.sequenceNumber++