@pgavlin 2023-05-12
>  I guess one possible anti-optimization here is for patterns that need a lot of memory early on and less memory at the end, like when deleting resources, it'll hold on to the memory longer in the buffer variable but that's likely worth it.

Yeah that was my feeling as well. If that proves to be a problem we could release the buffer if it gets too small.